feat: 북마크 조회 api JPA로 변환
