+++ src/main/java/com/kuit/conet/jpa/repository/MemberRepository.java
@@ -2,6 +2,7 @@ package com.kuit.conet.jpa.repository;
 
 import com.kuit.conet.dto.web.response.StorageImgResponseDTO;
 import com.kuit.conet.dto.web.response.team.GetTeamMemberResponseDTO;
+import com.kuit.conet.dto.web.response.team.GetTeamResponseDTO;
 import com.kuit.conet.jpa.domain.member.Member;
 import com.kuit.conet.jpa.domain.member.MemberStatus;
 import jakarta.persistence.EntityManager;
@@ -48,6 +49,13 @@ public class MemberRepository {
         return em.createQuery("select new com.kuit.conet.dto.web.response.team.GetTeamMemberResponseDTO(m.id,m.name,m.imgUrl) " +
                         "from TeamMember tm join tm.member m where tm.team.id=:teamId")
                 .setParameter("teamId", teamId)
+                .getResultList();
+    }
+
+    public List<GetTeamResponseDTO> getBookmarks(Long userId) {
+        return em.createQuery("select new com.kuit.conet.dto.web.response.team.GetTeamResponseDTO(t, tm.bookMark) " +
+                        "from TeamMember tm join tm.team t where tm.member.id=:userId", GetTeamResponseDTO.class)
+                .setParameter("userId", userId)
                 .getResultList();
     }
 }
